pyproject.toml
[project]
name = "inventory-manager"
version = "0.1.0"
description = "A desktop inventory management app with partners and suppliers"
authors = [
    { name = "Jowie", email = "jtanbp@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "greenlet==3.2.4",
    "psycopg2-binary==2.9.10",
    "PySide6==6.9.2",
    "PySide6-Addons==6.9.2",
    "PySide6-Essentials==6.9.2",
    "python-dotenv==1.1.1",
    "requests==2.32.5",
    "shiboken6==6.9.2",
    "typing_extensions==4.15.0"
]

[project.optional-dependencies]
dev = [
    "flake8==7.3.0",
    "flake8-isort==6.1.2",
    "isort==6.0.1",
    "mccabe==0.7.0",
    "pycodestyle==2.14.0",
    "pyflakes==3.4.0",
    "pytest==8.4.2",
    "pip-tools==7.5.0"  # lets you export requirements.txt
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

#[tool.flake8]
#max-line-length = 88
#extend-ignore = ["E203", "W503"]
#exclude = [".venv", "venv", "build", "dist", "__pycache__"]
#
#[tool.isort]
## Recognize your project packages as first-party
#known_first_party = ["core", "database", "ui"]
#
## Sections of imports: standard library, third-party, first-party, local
#sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
#
## Output formatting options
#multi_line_output = 3          # vertical hanging indent style
#include_trailing_comma = true  # trailing comma on multi-line imports
#force_grid_wrap = 0
#use_parentheses = true
#line_length = 88               # maximum line length
#
## Alphabetize imports within each section
#order_by_type = true
